requirements.txt
Django==1.6.5
python-social-auth==0.1.26
MySQL-python==1.2.5
django-debug-toolbar==1.2.1
django-extensions==1.3.10
django-autoslug==1.7.2
